refactor(worktree): curate barrel exports and fix direct-import callers
- Remove unused exports from worktree/index.ts: SelectedWorktreePlan,
  parseWorktreeList, isRalphaiManagedBranch
- Rewrite 5 direct imports from internal worktree files (management.ts,
  types.ts) to go through the barrel
- Update hitl.test.ts mock to match curated barrel surface
- Document curated public API in ARCHITECTURE.md
